cleaning warnings

+++ src/deps/md5/md5.h
@@ -25,6 +25,7 @@
 #ifndef _MD5_H_WRAPPER
 #define _MD5_H_WRAPPER
 
+#include <stdint.h>
 #if 1
 #include <openssl/md5.h>
 #elif !defined(_MD5_H)
@@ -44,11 +45,11 @@ typedef struct {
 } MD5_CTX;
 
 extern void MD5_Init(MD5_CTX *ctx);
-extern void MD5_Update(MD5_CTX *ctx, const void *data, unsigned long size);
+extern void MD5_Update(MD5_CTX *ctx, const void *data, uint32_t size);
 extern void MD5_Final(unsigned char *result, MD5_CTX *ctx);
 
 // openssl provide a function MD5(), we just wrap it out if the library is not used
-extern unsigned char * MD5( const unsigned char *pointer, unsigned long pointerLength, unsigned char *md5HashPointer);
+extern unsigned char * MD5( const unsigned char *pointer, uint32_t pointerLength, unsigned char *md5HashPointer);
 
 #endif
